Extract DSN class into its own file
