+++ spec/percona_migrator/cli_generator_spec.rb
@@ -120,14 +120,4 @@ describe PerconaMigrator::CliGenerator do
       it { is_expected.to include("D=#{connection_data[:database]}") }
     end
   end
-
-  describe PerconaMigrator::DSN do
-    let(:database) { 'development' }
-    let(:table_name) { 'comments' }
-
-    describe '#to_s' do
-      subject { described_class.new(database, table_name).to_s }
-      it { is_expected.to eq('D=development,t=comments') }
-    end
-  end
 end
